refactor: api 로직에 포함되어 있던 서버 데이터 fetching 관련 커스텀 훅을 별도 파일로 분리
