+++ src/apis/auth/socialSignIn.ts
@@ -1,8 +1,4 @@
-import { useQuery } from '@tanstack/react-query';
-
-import { USER } from '@/constants/API';
 import { SocialSignInAuthType } from '@/types/auth';
-import TokenHandler from '@/utils/tokenHandler';
 
 import defaultClient from '..';
 
@@ -12,25 +8,4 @@ const getSocialSignIn = async (queryUrl: string) => {
   return data;
 };
 
-const useGetSocialSignIn = ({
-  provider,
-  params,
-}: {
-  provider: string;
-  params: string;
-}) => {
-  const socialSignInUrl = `${USER.socialSignIn}/${provider.toUpperCase()}?${params}`;
-
-  const { data, isLoading } = useQuery({
-    queryKey: [`${USER.socialSignIn}/${provider}`],
-    queryFn: () => getSocialSignIn(socialSignInUrl),
-  });
-
-  if (data) {
-    TokenHandler.setToken(data.token);
-  }
-
-  return { hasData: !!data, isLoading };
-};
-
-export default useGetSocialSignIn;
+export default getSocialSignIn;
